+++ src/Query/Concerns/NonFilterable.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Katalam\OnOfficeAdapter\Query\Concerns;
+
+trait NonFilterable
+{
+    public function where(string $column, mixed $operator, mixed $value = null): self
+    {
+        return $this;
+    }
+}