+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## main
 - Moved user repository query to settings repository
 - Added regions to settings repository
+- Added non-filterable, non-limitable, non-sortable and non-selectable traits to repositories if needed
 
 ## v0.3.0
 - Added a fake implementation for the estate repository